part1.8
